Safari-609.1.11.1/LayoutTests/editing/selection/dump-as-markup-form-text.html

<!DocType html>
<script src="../../resources/dump-as-markup.js"></script>

<input value="original" />
<textarea>original</textarea>
<script>
Markup.description("Tests dump-as-markup for input and textarea forms.");

document.getElementsByTagName("input")[0].value = "replaced";
document.getElementsByTagName("textarea")[0].value = "replaced";
</script>
