Safari-609.1.11.1/LayoutTests/editing/selection/dump-as-markup-form-text-expected.txt

Tests dump-as-markup for input and textarea forms.
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "

"
|   <body>
|     <input>
|       value="original"
|       this.value="replaced"
|       <shadow:root>
|         <div>
|           contenteditable="plaintext-only"
|           "replaced"
|     "
"
|     <textarea>
|       this.value="replaced"
|       "original"
|       <shadow:root>
|         <div>
|           contenteditable="plaintext-only"
|           "replaced"
|     "
"
|     <script>
|       "
Markup.description("Tests dump-as-markup for input and textarea forms.");

document.getElementsByTagName("input")[0].value = "replaced";
document.getElementsByTagName("textarea")[0].value = "replaced";
"
|     "
"
